implement reading open & free formats + wider outfile format
tested with some CEP files around, and got to use wider G format
worked also with open format, but not tried with free format.

Failed with a data set where a site (lake) name was 'VITSJ\' (for
bad text encoding of Vitsjö), and \' was interpreted as escape
when source()ing the outfile. It may be that some other form than
source()able data.frame could be more robust against character
strings.