build only Win32 (i386) binary in Windows
There is no need to build separate binaries to Windows
sub-architectures because the binary and R communicate only
via system2() and file. I think i386 is more widely applicable
and we build only for that.

Makefile.win now includes R_HOME/etc/i386/Makeconf to set macros
for the i386 build. This gives a *long* NOTE in R CMD check. We
copy the model of CRAN package BRugs by Uwe Ligges which does the
same.