Windows build tools were updated for R-devel/R 4.2
Patch was supplied by Tomas Kalibera with following notes in email:

These patches are currently automatically applied by R-devel on Windows
at installation time, which makes most of your packages pass their checks
(as OK or NOTE), but please check your results carefully and carefully
review the patches. Usually these changes were because of newer GCC or
newer MinGW in the toolchain, but some for other reasons, and some of
them will definitely have to be improved so that the package keeps
building also for older versions of R using Rtools40.

In this case the fix seems only be that builds for i386 were
disabled – earlier I opted for i386 to guarantee working in older
Windows installation.